requirement.txt
alembic==0.9.6
amqp==2.2.2
aniso8601==1.3.0
anyjson==0.3.3
asn1crypto==0.23.0
astroid==1.5.3
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
blessings==1.6
blinker==1.4
bpython==0.16
celery==4.1.0
celery-with-redis==3.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.1.2
cssselect==1.0.1
curtsies==0.2.11
Django==1.11.6
django-celery==3.2.1
django-redis-sessions==0.6.1
django-tinymce==2.6.0
dominate==2.3.1
et-xmlfile==1.0.1
eyeD3==0.8.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap3==3.1.1.3
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SQLService==0.1.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
greenlet==0.4.12
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
kombu==4.1.0
lazy-object-proxy==1.3.1
lml==0.0.1
lxml==4.1.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.12
olefile==0.44
openpyxl==2.4.9
ordereddict==1.1
parsel==1.2.0
pathlib==1.0.1
pdfminer3k==1.3.1
Pillow==4.3.0
ply==3.10
py==1.4.34
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pydash==4.2.1
PyDispatcher==2.0.5
pyexcel==0.5.6
pyexcel-io==0.5.3
pyexcel-xls==0.5.2
pygame==1.9.3
Pygments==2.2.0
pylint==1.7.4
pymongo==3.5.1
PyMySQL==0.7.11
pyOpenSSL==17.3.0
pypyodbc==1.3.4
pytest==3.2.3
python-dateutil==2.6.1
python-editor==1.0.3
python-magic==0.4.13
pytz==2017.2
pywin32==221
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.1.14
sqlservice==0.14.2
texttable==1.1.1
tornado==4.5.2
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
visitor==0.1.3
w3lib==1.18.0
wcwidth==0.1.7
Werkzeug==0.12.2
wrapt==1.10.11
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
xlwt-future==0.8.0
zope.interface==4.4.3
